requirements.txt
flask
flask-socketio
flask-cors
flask-jwt-extended
eventlet
pymongo
datetime
typing
dataclasses
cloudinary
werkzeug
python-dotenv
cryptography